test: Move platform tests, remove qpu ones
