+++ tests/platform/dummy_hardware/platform.py
